Include unpublishable path dependencies in package
Rebase of: https://github.com/rust-lang/cargo/pull/4735

RFC: https://github.com/rust-lang/rfcs/pull/2224

Before this patch, all path deps must be published before the root crate
is published to the registry.

This patch allow `cargo package` to include files of path dependencies
in the generated package if the dependencies have the manifest key
"package.publish" set to false, which means the crate cannot be
registered as a sharable crate by itself at a registry, but still
allowed to be published along with the root crate.

Co-authored-by: J.W <jakwings@gmail.com>